Added focus, and recursive
